% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tern_xy.R
\name{tern_xy}
\alias{tern_xy}
\title{Convert a ternary triplet to Cartesian coordinates}
\usage{
tern_xy(a, b, c)
}
\arguments{
\item{a}{first}

\item{b}{second}

\item{c}{third}
}
\value{
a tibble with x/y parameters
}
\description{
Convert a ternary triplet to Cartesian coordinates
}
