% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternex.R
\docType{data}
\name{ternex}
\alias{ternex}
\title{Ternary Example data}
\format{
## `ternex`
A data frame with 60 rows and y columns:
\describe{
  \item{ID}{Individual ID}
  \item{A, B, C}{Proportion response in each of three categories, A, B, and C.}
  \item{Binary}{A binary variable (coded Yes/No)}
  \item{Trio}{A categorical variable with three levels (X, Y, Z)}
  \item{Quad}{A categorical variable with four levels (Q1, Q2, Q3, Q4)}
}
}
\usage{
ternex
}
\description{
An anonymized data set for use as an example of ternary plots.
}
\keyword{datasets}
