% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternit.R
\name{ternit}
\alias{ternit}
\title{Ternary Plot}
\usage{
ternit(
  data,
  labels,
  facets,
  grid.spacing = 0.2,
  tick.length = 0.03,
  tick.label.spacing = tick.length * 2,
  axis.label.spacing = tick.length * 4,
  lims
)
}
\arguments{
\item{data}{the data}

\item{labels}{axis labels}

\item{facets}{facets with data}

\item{grid.spacing}{grid spacing}

\item{tick.length}{tick length}

\item{tick.label.spacing}{tick label spacing}

\item{axis.label.spacing}{axis label spacing}

\item{lims}{}
}
\value{
a ternary plot
}
\description{
Ternary Plot
}
